Dispatch value on the main thread when binding from the main thread
When converting the publisher to a liveData instance, we were always dispatching the publisher values on the background thread even if we were always on the main thread. To fix the issue, we now use our own implementation of the LiveDataReactiveStreams which check the current thread before dispatching values. If we are already on the main thread, we use `setValue()` which dispatch the value synchronously and if we aren't on the main thread we use `postValue()` which post the value on the main thread.