Exponential backoff policy
I also added an extension to retry with backOff, it basically combines retryWhen, switchMap and timer. I was hesistant to add it here but it should be a good starting point.
For more advanced usage (retrying on certain errors, or only when network is available) the BackoffStrategy will be available standalone (it is actually my use case).